return props instead of component from hook
